requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
pydantic==2.10.4
python-dotenv==1.0.1
boto3==1.35.90
python-multipart==0.0.20
alembic==1.14.0
jinja2
reportlab
